type error
Traceback (most recent call last):
  File "main.py", line 64, in <module>
    main(config)
  File "main.py", line 33, in main
    solver.train()
  File "/home/user/PycharmProjects/SpeechSplit-master/solver.py", line 161, in train
    x_f0_intrp = self.Interp(x_f0, len_org)
  File "/home/user/anaconda3/envs/vencen/lib/python3.6/site-packages/torch/nn/modules/module.py", line 547, in __call__
    result = self.forward(*input, **kwargs)
  File "/home/user/PycharmProjects/SpeechSplit-master/model.py", line 413, in forward
    idx_scaled_org = idx_scaled_fl + offset
RuntimeError: expected device cuda:0 and dtype Float but got device cuda:0 and dtype Long
Aborted (core dumped)
exit status 134

+++ model.py
@@ -408,7 +408,7 @@ class InterpLnr(nn.Module):
         # offset starts from the 2nd segment
         offset = F.pad(offset[:, :-1], (1,0), value=0).view(-1, 1)
         
-        idx_scaled_org = idx_scaled_fl + offset
+        idx_scaled_org = idx_scaled_fl + offset.float()
         
         len_seq_rp = torch.repeat_interleave(len_seq, self.max_num_seg)
         idx_mask_org = idx_scaled_org < (len_seq_rp - 1).unsqueeze(-1)
